+++ CHANGELOG.md
@@ -2,7 +2,7 @@ CHANGELOG
 =========
 
 ## HEAD (Unreleased)
-* Upgrade to v2.2.1 of the NewRelic Terraform Provider
+* Upgrade to v2.3.0 of the NewRelic Terraform Provider
 * Upgrade to Pulumi v2.7.1
 * Upgrade to pulumi-terraform-bridge v2.6.0